crewai_software_engineer_workflow/requirements.txt
langchain-exa==0.0.1
langchain-community==0.0.38
langchain-groq==0.1.3
langchain-google-genai==1.0.4 
exa_py==1.0.7
python-dotenv==1.0.0
crewai==0.28.8
crewai-tools==0.2.6